==> data_cleaning/cleaning/1_delete_blank_and_uncertain_VDss.ipynb <==
import warnings
import pandas as pd

warnings.filterwarnings("ignore")

data = pd.read_excel(r"/opt/jupyter_data/data_clean/raw_data/1_all_data.xlsx")
print(data.shape)

def del_uncertain(df):
    """This function removes null values and transfers the range values to another table."""
    # delete empty VDss
    df = df.dropna(subset=['VDss_L_kg'])
    df["VDss_L_kg"] = df["VDss_L_kg"].astype(str)
    # remove data whose VDss is range
    df1 = df.loc[df["VDss_L_kg"].str.contains('-')]
    # delete VDss is the range of data
    df = df.loc[-df["VDss_L_kg"].str.contains('-')]
    df["VDss_L_kg"] = df["VDss_L_kg"].astype(float)
    
    return df, df1

if __name__ == "__main__":
    # separate VDss of exact value and range
    del_uncertain_data, range_data = del_uncertain(data)
    
    # save files
    del_uncertain_data.to_excel('2_delete_uncertain_VDss.xlsx', index=False)
    range_data.to_excel('1_range_VDss.xlsx', index=False)
