==> data_cleaning/cleaning/4_get_Isomeric_smiles.ipynb <==
import warnings
import numpy as np
import pandas as pd
from rdkit import Chem, RDLogger
from rdkit.Chem import MolToSmiles
from rdkit.Chem.inchi import MolToInchi
from pubchempy import get_compounds, Compound

warnings.filterwarnings("ignore")
RDLogger.DisableLog('rdApp.*')

data = pd.read_excel("4_delete_uncertain_smiles.xlsx")
print(data.shape)
data.head()

def rdkit_smiles(df):
    """This function checks and unifies SMILES."""  
    for smiles, i in zip(df["smiles_supplementary"], range(df.shape[0])):
        smi = 'error'
        # noinspection PyBroadException
        try:
            # unified SMILES format
            mol = Chem.MolFromSmiles(smiles)
            smi = Chem.MolToSmiles(mol)
            df.loc[i, 'isomeric_smiles'] = smi
        except:
            # regenerate SMILES with chemical name and CAS number
            # noinspection PyBroadException
            try:
                # use the CAS number or chemical name to
                # acquire isomeric SMILES to augment null SMILES
                cas = df.loc[i, 'cas_number']
                if pd.isnull(cas):
                    chem = df.loc[i, 'chemical_name']
                    if pd.isnull(chem):
                        df.loc[i, 'isomeric_smiles'] = smi
                    else:
                        # generate isomeric SMILES using the chemical name
                        for compound in get_compounds(chem, 'name'):
                            smiles = compound.isomeric_smiles
                        df.loc[i, 'smiles_supplementary'] = smiles
                else:
                    # generate isomeric SMILES using CAS number
                    for compound in get_compounds(cas, 'name'):
                        smiles = compound.isomeric_smiles
                    df.loc[i, 'smiles_supplementary'] = smiles
                # unify SMILES
                mol = Chem.MolFromSmiles(smiles)
                smi = Chem.MolToSmiles(mol)
                df.loc[i, 'isomeric_smiles'] = smi
            except:
                df.loc[i, 'isomeric_smiles'] = smi

    return df

def modify_smiles(df):
    """This function gets correct RDKit SMILES."""
    smiles = 'error'
    for chem, i in zip(df['chemical_name'], range(df.shape[0])):
        # noinspection PyBroadException
        try:
            if pd.isnull(chem):
                continue
            else:
                # generate isomeric SMILES using the chemical name
                for compound in get_compounds(chem, 'name'):
                    smiles = compound.isomeric_smiles
                df.loc[i, 'smiles_supplementary'] = smiles
                # get RDKit SMILES
                mol = Chem.MolFromSmiles(smiles)
                smi = Chem.MolToSmiles(mol)
                df.loc[i, 'isomeric_smiles'] = smi
        except:
            continue

    return df

if __name__ == "__main__":
    # check SMILES
    data = rdkit_smiles(data)
    # filter out data that contains 'error'
    error = data.loc[data["isomeric_smiles"].str.contains('error')]
    # update index
    error.index = range(error.shape[0])
    # update the wrong SMILES
    modify_error = modify_smiles(error)
    # combine data
    data = pd.concat([modify_error, data], ignore_index=True)

# delete the row that contains the content of 'error' in the column 'isomeric_smiles'
data = data.loc[-data["isomeric_smiles"].str.contains('error')]
# update index
data.index = range(data.shape[0])
# output data shape
print(data.shape)

# save file
data.to_excel('5_delete_rdkit_smiles_error.xlsx', index=False)
